Rename "redux" folder to "store" and update imports
