+++ src/store/boardSlice.ts
